complete table

+++ modules/hivemq-edge-module-opcua/src/main/java/com/hivemq/edge/adapters/opcua/writing/JsonToOpcUAConverter.java
@@ -283,7 +283,7 @@ public class JsonToOpcUAConverter {
                 throw createException(jsonNode, builtinDataType.name());
             case DateTime:
                 if (jsonNode.isLong()) {
-                    return new DateTime(jsonNode.asInt());
+                    return new DateTime(jsonNode.asLong());
                 }
                 throw createException(jsonNode, builtinDataType.name());
             case Guid:
